+++ lib/minfraud/error_handler.rb
@@ -27,15 +27,6 @@ module Minfraud
 
       # @!visibility private
       STATUS_CODES = {
-        IP_ADDRESS_INVALID:    [
-          ClientError, 'You have not supplied a valid IPv4 or IPv6 address'
-        ],
-        IP_ADDRESS_REQUIRED:   [
-          ClientError, 'You have not supplied an IP address which is a required field'
-        ],
-        IP_ADDRESS_RESERVED:   [
-          ClientError, 'You have supplied an IP address which is reserved'
-        ],
         JSON_INVALID:          [
           ClientError, 'JSON body cannot be decoded'
         ],
@@ -47,6 +38,9 @@ module Minfraud
         ],
         PARAMETER_UNKNOWN:     [
           ClientError, 'You have supplied an unknown parameter'
+        ],
+        REQUEST_INVALID:       [
+          ClientError, 'The request did not contain any valid input values.'
         ],
         TAG_REQUIRED:          [
           ClientError, 'You have not supplied a tag, which is a required field'