+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * IMPORTANT: Ruby 2.0 is no longer supported. If you're using Ruby 2.0,
   please use version 1.3.0.
+* Add handling for the `REQUEST_INVALID` error code.
+* The IP address is no longer a required input.
 
 ## v1.3.0 (2020-09-25)